WLA-442: Move LaunchDarkly to middleware
